refactor: move earn components
