+++ packages/web/src/components/pool/position-history-table/PositionHistoryTable.tsx
@@ -7,16 +7,18 @@ import {
   MOBILE_POSITION_HISTORY_INFO,
   POSITION_HISTORY_INFO,
   TableInfoType,
-  TABLET_POSITION_HISTORY_INFO
+  TABLET_POSITION_HISTORY_INFO,
 } from "@constants/skeleton.constant";
 import { TABLE_HEAD } from "@containers/position-history-container/PositionHistoryContainer";
 import { IPositionHistoryModel } from "@models/position/position-history-model";
 import { DEVICE_TYPE } from "@styles/media";
 
 import PositionInfo from "../position-info/PositionInfo";
+
 import {
-  noDataText, TableHeader,
-  TableWrapper
+  noDataText,
+  TableHeader,
+  TableWrapper,
 } from "./PositionHistoryTable.styles";
 
 interface PositionHistoryTableProps {
