+++ packages/web/src/components/wallet/asset-list-table/AssetListTable.tsx
@@ -1,23 +1,28 @@
-import {
-  type Asset,
-  ASSET_HEAD,
-  AssetSortOption,
-} from "@containers/asset-list-container/AssetListContainer";
-import AssetInfo from "@components/wallet/asset-info/AssetInfo";
-import { AssetListTableWrapper, TableColumn } from "./AssetListTable.styles";
-import { noDataText } from "@components/earn/pool-list-table/PoolListTable.styles";
+import { cx } from "@emotion/css";
+import { useCallback, useMemo } from "react";
+import { useTranslation } from "react-i18next";
+
+import IconTriangleArrowDown from "@components/common/icons/IconTriangleArrowDown";
+import IconTriangleArrowUp from "@components/common/icons/IconTriangleArrowUp";
 import TableSkeleton from "@components/common/table-skeleton/TableSkeleton";
+import AssetInfo from "@components/wallet/asset-info/AssetInfo";
 import {
   ASSET_INFO,
   ASSET_INFO_MOBILE,
   ASSET_INFO_TABLET,
 } from "@constants/skeleton.constant";
-import { cx } from "@emotion/css";
-import { useCallback, useMemo } from "react";
-import IconTriangleArrowDown from "@components/common/icons/IconTriangleArrowDown";
-import IconTriangleArrowUp from "@components/common/icons/IconTriangleArrowUp";
+import {
+  AssetSortOption,
+  ASSET_HEAD,
+  type Asset,
+} from "@containers/asset-list-container/AssetListContainer";
 import { DEVICE_TYPE } from "@styles/media";
-import { useTranslation } from "react-i18next";
+
+import {
+  AssetListTableWrapper,
+  noDataText,
+  TableColumn,
+} from "./AssetListTable.styles";
 
 interface AssetListTableProps {
   assets: Asset[];
