+++ packages/web/src/views/earn/components/pool-list/PoolList.stories.tsx
@@ -2,11 +2,12 @@ import React from "react";
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 
+import { DEVICE_TYPE } from "@styles/media";
+
 import PoolList from "./PoolList";
 import {
   POOL_TYPE,
-} from "@containers/pool-list-container/PoolListContainer";
-import { DEVICE_TYPE } from "@styles/media";
+} from "./types";
 
 export default {
   title: "earn/PoolList",
