+++ packages/web/src/views/earn/components/pool-list/pool-list-table/pool-info/PoolInfo.tsx
@@ -1,21 +1,24 @@
-import DoubleLogo from "@components/common/double-logo/DoubleLogo";
 import React, { useMemo } from "react";
-import { PoolInfoWrapper, TableColumn } from "./PoolInfo.styles";
-import { PoolListInfo } from "@models/pool/info/pool-list-info";
+
+import DoubleLogo from "@components/common/double-logo/DoubleLogo";
+import IconStar from "@components/common/icons/IconStar";
+import OverlapTokenLogo from "@components/common/overlap-token-logo/OverlapTokenLogo";
 import { SwapFeeTierInfoMap } from "@constants/option.constant";
-import { useGnotToGnot } from "@hooks/token/use-gnot-wugnot";
-import { DEVICE_TYPE } from "@styles/media";
-import PoolInfoLazyChart from "../pool-info-lazy-chart/PoolInfoLazyChart";
 import {
   POOL_INFO,
   POOL_INFO_MOBILE,
   POOL_INFO_SMALL_TABLET,
-  POOL_INFO_TABLET,
+  POOL_INFO_TABLET
 } from "@constants/skeleton.constant";
-import IconStar from "@components/common/icons/IconStar";
-import OverlapTokenLogo from "@components/common/overlap-token-logo/OverlapTokenLogo";
+import { useGnotToGnot } from "@hooks/token/use-gnot-wugnot";
+import { PoolListInfo } from "@models/pool/info/pool-list-info";
 import { TokenModel } from "@models/token/token-model";
+import { DEVICE_TYPE } from "@styles/media";
 import { formatRate } from "@utils/new-number-utils";
+
+import PoolInfoLazyChart from "./pool-info-lazy-chart/PoolInfoLazyChart";
+
+import { PoolInfoWrapper, TableColumn } from "./PoolInfo.styles";
 
 interface PoolInfoProps {
   pool: PoolListInfo;
