+++ packages/web/src/views/earn/components/pool-list/pool-list-table/pool-info/pool-info-lazy-chart/PoolInfoLazyChart.styles.ts
